+++ waveorder/io/writer_structures.py
@@ -1,8 +1,9 @@
 import os
 from numcodecs import Blosc
 import numpy as np
+import zarr
 
-class Builder:
+class WriterBase:
     """
     ABC for all builders
     """
@@ -92,6 +93,24 @@ class Builder:
         else:
             raise ValueError('Did not understand data formatting')
 
+    def create_channel_dict(self, chan_name, clim=None):
+
+        min = 0.0
+        max = 65535.0
+        start = clim[0] if clim else 0.0
+        end = clim[1] if clim else 65535.0
+
+        dict_ = {'active': True,
+                'coefficient': 1.0,
+                'color': '808080',
+                'family': 'linear',
+                'inverted': False,
+                'label': chan_name,
+                'window': {'end': end, 'max': max, 'min': min, 'start': start}
+                }
+
+        return dict_
+
     def set_channel_attributes(self, chan_names: list, clims: list):
         """
         A method for creating ome-zarr metadata dictionary.
@@ -155,11 +174,152 @@ class Builder:
             print(f'Found existing store at {path}')
             return True
         else:
-            # print(f'Creating new store at {path}')
             return False
 
-    def create_channel_dict(self, chan_names, clims):
-        pass
 
-    def init_compressor(self):
-        pass
+class PhysicalZarr(WriterBase):
+
+    def __init__(self, name=None):
+        """
+
+        """
+        super().__init__()
+        self.__zarr = None
+
+        if name:
+            self.name = name
+        else:
+            self.name = 'physical_data'
+
+    def create_channel_dict(self, chan_name, clim=None):
+
+        if chan_name == 'Retardance':
+            min = 0.0
+            max = 1000.0
+            start = clim[0] if clim else 0.0
+            end = clim[1] if clim else 100.0
+        elif chan_name == 'Orientation':
+            min = 0.0
+            max = np.pi
+            start = clim[0] if clim else 0.0
+            end = clim[1] if clim else np.pi
+
+        elif chan_name == 'Phase3D':
+            min = -10.0
+            max = 10.0
+            start = clim[0] if clim else -0.2
+            end = clim[1] if clim else 0.2
+
+        elif chan_name == 'BF':
+            min = 0.0
+            max = 65535.0
+            start = clim[0] if clim else 0.0
+            end = clim[1] if clim else 5.0
+
+        else:
+            min = 0.0
+            max = 65535.0
+            start = clim[0] if clim else 0.0
+            end = clim[1] if clim else 65535.0
+
+        dict_ = {'active': True,
+                'coefficient': 1.0,
+                'color': '808080',
+                'family': 'linear',
+                'inverted': False,
+                'label': chan_name,
+                'window': {'end': end, 'max': max, 'min': min, 'start': start}
+                }
+
+        return dict_
+
+
+class StokesZarr(WriterBase):
+
+    def __init__(self, name=None):
+        """
+
+        """
+        super().__init__()
+        self.__zarr = None
+
+        if name:
+            self.name = name
+        else:
+            self.name = 'stokes_data'
+
+    def create_channel_dict(self, chan_name, clim=None):
+
+        if chan_name == 'S0':
+            min = 0.0
+            max = 65535
+            start = clim[0] if clim else 0.0
+            end = clim[1] if clim else 1.0
+
+        elif chan_name == 'S1':
+            min = 10.0
+            max = -10.0
+            start = clim[0] if clim else -0.5
+            end = clim[1] if clim else 0.5
+
+        elif chan_name == 'S2':
+            min = -10.0
+            max = 10.0
+            start = clim[0] if clim else -0.5
+            end = clim[1] if clim else 0.5
+
+        elif chan_name == 'S3':
+            min = -10
+            max = 10
+            start = clim[0] if clim else -1.0
+            end = clim[1] if clim else 1.0
+
+        else:
+            min = 0.0
+            max = 65535.0
+            start = clim[0] if clim else 0.0
+            end = clim[1] if clim else 65535.0
+
+        dict_ = {'active': True,
+                'coefficient': 1.0,
+                'color': '808080',
+                'family': 'linear',
+                'inverted': False,
+                'label': chan_name,
+                'window': {'end': end, 'max': max, 'min': min, 'start': start}
+                }
+
+        return dict_
+
+
+class RawZarr(WriterBase):
+
+    def __init__(self, name=None):
+        """
+
+        """
+        super().__init__()
+        self.__zarr = None
+
+        if name:
+            self.name = name
+        else:
+            self.name = 'raw_data'
+
+    def create_channel_dict(self, chan_name, clim=None):
+
+        min = 0
+        max = 65535
+        start = clim[0] if clim else 0
+        end = clim[1] if clim else 65535
+
+        dict_ = {'active': True,
+                'coefficient': 1.0,
+                'color': '808080',
+                'family': 'linear',
+                'inverted': False,
+                'label': chan_name,
+                'window': {'end': end, 'max': max, 'min': min, 'start': start}
+                }
+
+        return dict_
